resolves #367 by saving carpool pickup search seed
we add a column to the carpool table called "from_seed". this field stores the
search string the user entered to find their dropoff point in the new carpool
form. we persist this so that the same location is available in that field and
on the map below when they navigate to the edit form